fixed entity, .properiest converted to .yaml

+++ src/main/java/ru/emiljan/servicedevdevices/models/Image.java
@@ -3,7 +3,6 @@ package ru.emiljan.servicedevdevices.models;
 import lombok.*;
 
 import javax.persistence.*;
-import java.util.List;
 
 /**
  * @author EM1LJAN
@@ -18,7 +17,7 @@ import java.util.List;
 public class Image {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
 
     @Column(name = "icon64")
     private String icon64;
